tab links and urls on main page
