+++ web/src/App.js
@@ -1,6 +1,5 @@
 import './App.css';
 import React from 'react';
-import gds from './api/gds';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Tab from 'react-bootstrap/Tab';
@@ -8,24 +7,10 @@ import Tabs from 'react-bootstrap/Tabs';
 import Lookup from './components/Lookup';
 import Alerts from './components/Alerts';
 import Registration from './components/Registration';
-import LookupResults from './components/LookupResults';
 
 
 class App extends React.Component {
-  state = { results: {}, alerts: [] };
-
-  onLookup = async (query, inputType) => {
-    try {
-      const response = await gds.lookup(query, inputType)
-      this.setState({ results: response })
-    } catch(err) {
-      this.setState(prevState => ({
-        results: {},
-        alerts: [...prevState.alerts, {variant: 'danger', message: err.message}]
-      }));
-      console.warn(err);
-    }
-  }
+  state = { alerts: [], currentPath: window.location.pathname };
 
   onAlert = (variant, message) => {
     this.setState(prevState => ({
@@ -37,6 +22,11 @@ class App extends React.Component {
     let alerts = [...this.state.alerts];
     alerts.splice(idx, 1);
     this.setState({alerts: alerts});
+  }
+
+  onTabSelect = (key) => {
+    window.history.pushState({}, '', key);
+    this.setState({ currentPath: window.location.pathname });
   }
 
   render() {
@@ -55,20 +45,11 @@ class App extends React.Component {
           </Col>
         </Row>
 
-        <Tabs defaultActiveKey="directory" id="tab-nav" className="justify-content-center">
-          <Tab eventKey="directory" title="Directory">
-            <Row className="py-3">
-              <Col md={{span: 8, offset: 2}}>
-                <Lookup onSubmit={this.onLookup} />
-              </Col>
-            </Row>
-            <Row>
-              <Col>
-                <LookupResults results={this.state.results} />
-              </Col>
-            </Row>
+        <Tabs activeKey={this.state.currentPath} id="main-tab-nav" className="justify-content-center" onSelect={this.onTabSelect}>
+          <Tab eventKey="/" title="Directory">
+            <Lookup onAlert={this.onAlert} />
           </Tab>
-          <Tab eventKey="register" title="Register">
+          <Tab eventKey="/register" title="Register">
             <Registration onAlert={this.onAlert} />
           </Tab>
         </Tabs>
